requirements.txt
allure-pytest==2.8.32
allure-python-commons==2.8.32
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
decorator==4.4.2
Django==3.1.1
idna==2.10
importlib-metadata==3.4.0
jsonpath-rw==1.4.0
more-itertools==8.6.0
pluggy==0.13.1
ply==3.11
py==1.10.0
PyMySQL==0.10.0
pytest==4.5.0
pytz==2020.1
PyYAML==5.4.1
requests==2.24.0
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
yaml-1.3==0.1.0
zipp==3.4.0
